+++ tests/test_rule_cache.py
@@ -6,10 +6,13 @@
 #  is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
 
+import os
 import textwrap
 import contextlib
+from pathlib import Path
 
 import capa.rules
+import capa.helpers
 import capa.rules.cache
 
 R1 = capa.rules.Rule.from_yaml(
@@ -113,3 +116,40 @@ def test_ruleset_cache_invalid():
     assert capa.rules.cache.load_cached_ruleset(cache_dir, content) is None
     # the invalid cache should be deleted
     assert not path.exists()
+
+
+def test_rule_cache_dev_environment():
+    # generate rules cache
+    rs = capa.rules.RuleSet([R2])
+    content = capa.rules.cache.get_ruleset_content(rs)
+    id = capa.rules.cache.compute_cache_identifier(content)
+    cache_dir = capa.rules.cache.get_default_cache_directory()
+    cache_path = capa.rules.cache.get_cache_path(cache_dir, id)
+
+    # clear existing cache files
+    for f in cache_dir.glob("*.cache"):
+        f.unlink()
+
+    capa.rules.cache.cache_ruleset(cache_dir, rs)
+    assert cache_path.exists()
+
+    assert capa.helpers.is_cache_newer_than_rule_code(cache_dir) is True
+
+    capa_root = Path(__file__).resolve().parent.parent
+    cachepy = capa_root / "capa" / "rules" / "cache.py"  # alternative: capa_root / "capa" / "rules" / "__init__.py"
+
+    # set cache's last modified time prior to code file's modified time
+    os.utime(cache_path, (cache_path.stat().st_atime, cachepy.stat().st_mtime - 600000))
+
+    # debug
+    def ts_to_str(ts):
+        from datetime import datetime
+
+        return datetime.fromtimestamp(ts).strftime("%Y-%m-%d %H:%M:%S")
+
+    for g in ((capa_root / "capa" / "rules").glob("*.py"), cache_dir.glob("*.cache")):
+        for p in g:
+            print(p, "\t", ts_to_str(p.stat().st_mtime))  # noqa: T201
+
+    assert capa.helpers.is_dev_environment() is True
+    assert capa.helpers.is_cache_newer_than_rule_code(cache_dir) is False